assets: define overridable components in SearchApp
