+++ invenio_search_ui/assets/semantic-ui/js/invenio_search_ui/components/SearchBar.js
