+++ invenio_search_ui/assets/semantic-ui/js/invenio_search_ui/components/index.js
@@ -1,0 +1,14 @@
+/*
+ * This file is part of Invenio.
+ * Copyright (C) 2020 CERN.
+ *
+ * Invenio is free software; you can redistribute it and/or modify it
+ * under the terms of the MIT License; see LICENSE file for more details.
+ */
+
+export { ResultsGridItem } from "./ResultsGridItem";
+export { ResultsListItem } from "./ResultsListItem";
+export { Results } from "./Results";
+export { SearchBar } from "./SearchBar";
+export { SearchApp, SearchAppFacets, SearchAppResultsPane } from "./SearchApp";
+export { SearchConfigurationContext } from "./context";
